+++ site/components/sandbox.tsx
@@ -2,12 +2,12 @@ import * as Solid from 'solid-js';
 import examples, { type ExampleModule } from '@app/example';
 import Fallback from '@app/fallback';
 import { isFunction } from '@moneko/common';
+import CodeLive, { type CodeLiveProps } from 'n-code-live';
 import { customElement } from 'solid-element';
 import h from 'solid-js/h';
 import * as SolidWeb from 'solid-js/web';
 import { codeNoShadowCss, groupCss, mdNoShadowCss, sandboxCss } from './sandbox.style';
 import * as NekoUI from 'neko-ui';
-import type { CodeLiveProps } from 'n-code-live';
 
 const { createEffect, createMemo, createSignal, mergeProps, onMount } = Solid;
 const { For, Show, render, Portal, Dynamic } = SolidWeb;
@@ -29,6 +29,7 @@ interface SandboxProps extends Omit<ExampleModule, 'title'> {
 
 const components: CodeLiveProps['components'] = {
   ...Solid,
+  CodeLive,
   NekoUI,
   Portal,
   Dynamic,